requirements.txt
# ──────────────────────────────────────────────────
# TrueLens AI — Python Dependencies
# ──────────────────────────────────────────────────

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Forensics
opencv-python>=4.8.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
